requirements.txt
streamlit
pulp
pandas
numpy
matplotlib.pyplot